Modified code that sub two unsigned ints and abs them
so, the warning was "taking the absolute value of unsigned type 'unsigned int' has no effect" so a little more info on that.

Although it says that imagine the case where x - y = a negative number, and we want to get the abs of that..
I dont know why that is unheard of of the compiler. but in the end, it does some crazy math to solve subtraction of two unsigneds.
check out http://stackoverflow.com/questions/7221409/is-unsigned-integer-subtraction-defined-behavior for more info. the clock example is good