Commented out some wierd pointer stuff
See what happens here is, strings are stored as a pointer to them inside a structure.
These structures are initialized at compile time yet, in 64 bit mode, these types of strings are not compile time constants
so there is no memory address. Playing with a small program i noticed that strings do not have a deterministic address.
Maybe this is why it will not allow us to cast one of these strings to an int... or better yet. to a long for 64 bit.