Updated struct definition to conform for 64bit operation
Basically the layout of the struct contains a pointer. This struct corresponds with a layout in WAD files that must be maintained. Unfortunately a pointer bak in the day was 32 bit and played nicely with the info in the wad. now its 64bit which causes the next two pieces of data to be offset by 4 bytes. In this case, we are lucky, because the void** is obsolete. So i changed it to an int and we are  now good to go. We do not have to do any pointer math or anything. we just use the fact that we dont need it and a int is 4 bytes.