Added comments and changed a little code
HERE IS THE WOA - basically there was a 4 hard coded into the mallocs

One - they like multiples of 4, so i can see why the 4 would  be there.
Two - 4 used to be a pointer size, which is what most of these are. textures, texturecolumnlump -- texturecomposite are all chunks of memory that store pointers
Three - although shorts are sized 2 (textureheight), they still want the  multiple of 4
So i updated the codez to be more dynamic
NEW
    textures = Z_Malloc(numtextures * sizeof(texture_t*), PU_STATIC, 0);
    texturecolumnlump = Z_Malloc(numtextures * sizeof(short*), PU_STATIC, 0);
    texturecolumnofs = Z_Malloc(numtextures * sizeof(unsigned short*), PU_STATIC, 0);
    texturecomposite = Z_Malloc(numtextures * sizeof(byte*), PU_STATIC, 0);
    texturecompositesize = Z_Malloc(numtextures * sizeof(int), PU_STATIC, 0);
    texturewidthmask = Z_Malloc(numtextures * sizeof(int), PU_STATIC, 0);
    textureheight = Z_Malloc(numtextures * sizeof(fixed_t), PU_STATIC, 0);

OLD
    textures = Z_Malloc(numtextures * 4, PU_STATIC, 0);
    texturecolumnlump = Z_Malloc(numtextures * 4, PU_STATIC, 0);
    texturecolumnofs = Z_Malloc(numtextures * 4, PU_STATIC, 0);
    texturecomposite = Z_Malloc(numtextures * 4, PU_STATIC, 0);
    texturecompositesize = Z_Malloc(numtextures * 4, PU_STATIC, 0);
    texturewidthmask = Z_Malloc(numtextures * 4, PU_STATIC, 0);
    textureheight = Z_Malloc(numtextures * 4, PU_STATIC, 0);